+++ L06 Workspace/Acme-HandyWorker/src/main/java/controllers/SocialProfileRefereeController.java
@@ -24,6 +24,7 @@ import org.springframework.web.servlet.ModelAndView;
 import security.LoginService;
 import services.RefereeService;
 import services.SocialProfilelService;
+import services.WelcomeService;
 import domain.Referee;
 import domain.SocialProfile;
 
@@ -35,6 +36,8 @@ public class SocialProfileRefereeController extends AbstractController {
 	private SocialProfilelService	socialProfileService;
 	@Autowired
 	private RefereeService			refereeService;
+	@Autowired
+	private WelcomeService			welcomeService;
 
 
 	//	@Autowired
@@ -54,6 +57,10 @@ public class SocialProfileRefereeController extends AbstractController {
 		socialProfile = this.socialProfileService.create();
 
 		result = new ModelAndView("socialProfile/referee/create");
+		final String system = this.welcomeService.getSystem();
+		result.addObject("system", system);
+		final String logo = this.welcomeService.getLogo();
+		result.addObject("logo", logo);
 
 		result.addObject("socialProfile", socialProfile);
 
@@ -61,17 +68,23 @@ public class SocialProfileRefereeController extends AbstractController {
 	}
 
 	@RequestMapping(value = "/edit", method = RequestMethod.GET)
-	public ModelAndView edit(@RequestParam("socialProfileId") final int socialProfileId) {
+	public ModelAndView edit(@RequestParam(value = "socialProfileId", defaultValue = "-1") final int socialProfileId) {
 		ModelAndView result;
+		if (this.socialProfileService.findOne(socialProfileId) == null || !this.refereeService.findByUserAccountId(LoginService.getPrincipal().getId()).getSocialProfiles().contains(this.socialProfileService.findOne(socialProfileId)))
+			result = new ModelAndView("welcome/index");
+		else {
+			final SocialProfile socialProfile;
+			Assert.isTrue(this.socialProfileService.findOne(socialProfileId) != null);
+			socialProfile = this.socialProfileService.findOne(socialProfileId);
 
-		final SocialProfile socialProfile;
-		Assert.isTrue(this.socialProfileService.findOne(socialProfileId) != null);
-		socialProfile = this.socialProfileService.findOne(socialProfileId);
+			result = new ModelAndView("socialProfile/referee/edit");
+			final String system = this.welcomeService.getSystem();
+			result.addObject("system", system);
+			final String logo = this.welcomeService.getLogo();
+			result.addObject("logo", logo);
 
-		result = new ModelAndView("socialProfile/referee/edit");
-
-		result.addObject("socialProfile", socialProfile);
-
+			result.addObject("socialProfile", socialProfile);
+		}
 		return result;
 	}
 
@@ -98,6 +111,10 @@ public class SocialProfileRefereeController extends AbstractController {
 				final Referee savedreferee = this.refereeService.save(referee);
 
 				result = new ModelAndView("referee/show");
+				final String system = this.welcomeService.getSystem();
+				result.addObject("system", system);
+				final String logo = this.welcomeService.getLogo();
+				result.addObject("logo", logo);
 				result.addObject("referee", savedreferee);
 				result.addObject("socialProfiles", savedreferee.getSocialProfiles());
 				result.addObject("requestURI", "referee/show.do");
@@ -111,28 +128,46 @@ public class SocialProfileRefereeController extends AbstractController {
 	}
 
 	@RequestMapping(value = "/delete", method = RequestMethod.GET)
-	public ModelAndView delete(@RequestParam("socialProfileId") final int socialProfileId) {
+	public ModelAndView delete(@RequestParam(value = "socialProfileId", defaultValue = "-1") final int socialProfileId) {
 		ModelAndView result;
-
-		final SocialProfile socialProfile = this.socialProfileService.findOne(socialProfileId);
-		System.out.println("socialProfileId encontrado: " + socialProfileId);
-		Assert.notNull(socialProfileId, "socialProfile.null");
-
-		try {
+		if (this.socialProfileService.findOne(socialProfileId) == null || !this.refereeService.findByUserAccountId(LoginService.getPrincipal().getId()).getSocialProfiles().contains(this.socialProfileService.findOne(socialProfileId))) {
+			System.out.println("socialProfileId encontrado: " + socialProfileId);
 			final int userLoggin = LoginService.getPrincipal().getId();
 			final Referee referee = this.refereeService.findByUserAccountId(userLoggin);
 			Assert.isTrue(referee != null);
-			referee.getSocialProfiles().remove(socialProfile);
-			this.refereeService.save(referee);
-			this.socialProfileService.delete(socialProfile);
 			result = new ModelAndView("referee/show");
 			result.addObject("referee", referee);
+			final String system = this.welcomeService.getSystem();
+			result.addObject("system", system);
+			final String logo = this.welcomeService.getLogo();
+			result.addObject("logo", logo);
 			result.addObject("socialProfiles", referee.getSocialProfiles());
 			result.addObject("requestURI", "referee/show.do");
-		} catch (final Throwable oops) {
-			System.out.println("Error al borrar socialProfile desde hw: ");
-			System.out.println(oops);
-			result = new ModelAndView("referee/show");
+		} else {
+			final SocialProfile socialProfile = this.socialProfileService.findOne(socialProfileId);
+			System.out.println("socialProfileId encontrado: " + socialProfileId);
+			Assert.notNull(socialProfileId, "socialProfile.null");
+
+			try {
+				final int userLoggin = LoginService.getPrincipal().getId();
+				final Referee referee = this.refereeService.findByUserAccountId(userLoggin);
+				Assert.isTrue(referee != null);
+				referee.getSocialProfiles().remove(socialProfile);
+				this.refereeService.save(referee);
+				this.socialProfileService.delete(socialProfile);
+				result = new ModelAndView("referee/show");
+				result.addObject("referee", referee);
+				final String system = this.welcomeService.getSystem();
+				result.addObject("system", system);
+				final String logo = this.welcomeService.getLogo();
+				result.addObject("logo", logo);
+				result.addObject("socialProfiles", referee.getSocialProfiles());
+				result.addObject("requestURI", "referee/show.do");
+			} catch (final Throwable oops) {
+				System.out.println("Error al borrar socialProfile desde hw: ");
+				System.out.println(oops);
+				result = new ModelAndView("referee/show");
+			}
 		}
 		return result;
 	}
